Update field.class.php
since  http://digitalbush.com/projects/masked-input-plugin/ is no longer accessible, I have added a simple javascript function to perform the mask the field and be sure that the inserted value is already formatted inside the moodle, this allows the entry always be inserted with 999.999.999-99 format.
This mask javascript is self-contained and independent of any external library. Also tested in all current browsers.